Repository 구현체에서 Entity를 Item으로 변경
