+++ feature-jinkwang/src/main/kotlin/com/samchi/poke/feature/jinkwang/JinKwangViewModel.kt
@@ -3,12 +3,9 @@ package com.samchi.poke.feature.jinkwang
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.paging.cachedIn
-import androidx.paging.map
 import com.samchi.poke.feature.jinkwang.data.JinKwangRepository
 import com.samchi.poke.feature.jinkwang.data.Pokemon
-import com.samchi.poke.feature.jinkwang.data.local.pokemon.PokemonEntity
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
@@ -18,11 +15,7 @@ internal class JinKwangViewModel @Inject constructor(
 ) : ViewModel() {
 
     val pokemonPaging = jinKwangRepository.getPokemonFlow()
-        .map { pokemonPaging ->
-            pokemonPaging.map { entity ->
-                entity.toItem()
-            }
-        }.cachedIn(viewModelScope)
+        .cachedIn(viewModelScope)
 
     fun retry() {
 
@@ -36,13 +29,5 @@ internal class JinKwangViewModel @Inject constructor(
                 jinKwangRepository.favoritePokemon(pokemon.nameField)
             }
         }
-    }
-
-    private fun PokemonEntity.toItem(): Pokemon {
-        return Pokemon(
-            nameField = name,
-            imageUrl = url,
-            isFavorite = isFavorite,
-        )
     }
 }
